+++ operator/internal/manifests/node_placement_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	lokiv1beta1 "github.com/grafana/loki/operator/api/v1beta1"
+	"github.com/grafana/loki/operator/internal/manifests/storage"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 )
@@ -44,7 +45,7 @@ func TestTolerationsAreSetForEachComponent(t *testing.T) {
 				},
 			},
 		},
-		ObjectStorage: ObjectStorage{},
+		ObjectStorage: storage.Options{},
 	}
 
 	optsWithoutTolerations := Options{
@@ -70,7 +71,7 @@ func TestTolerationsAreSetForEachComponent(t *testing.T) {
 				},
 			},
 		},
-		ObjectStorage: ObjectStorage{},
+		ObjectStorage: storage.Options{},
 	}
 
 	t.Run("distributor", func(t *testing.T) {
@@ -135,7 +136,7 @@ func TestNodeSelectorsAreSetForEachComponent(t *testing.T) {
 				},
 			},
 		},
-		ObjectStorage: ObjectStorage{},
+		ObjectStorage: storage.Options{},
 	}
 
 	optsWithoutNodeSelectors := Options{
@@ -161,7 +162,7 @@ func TestNodeSelectorsAreSetForEachComponent(t *testing.T) {
 				},
 			},
 		},
-		ObjectStorage: ObjectStorage{},
+		ObjectStorage: storage.Options{},
 	}
 
 	t.Run("distributor", func(t *testing.T) {